+++ test/unit/clitests_helper.js
@@ -15,8 +15,6 @@
 
 import { setVerbosityLevel, VerbosityLevel } from "../../src/shared/util.js";
 import { isNodeJS } from "../../src/shared/is_node.js";
-import { PDFNodeStream } from "../../src/display/node_stream.js";
-import { setPDFNetworkStreamFactory } from "../../src/display/api.js";
 
 // Sets longer timeout, similar to `jasmine-boot.js`.
 jasmine.DEFAULT_TIMEOUT_INTERVAL = 30000;
@@ -31,8 +29,3 @@ if (!isNodeJS) {
 // Reduce the amount of console "spam", by ignoring `info`/`warn` calls,
 // when running the unit-tests in Node.js/Travis.
 setVerbosityLevel(VerbosityLevel.ERRORS);
-
-// Set the network stream factory for the unit-tests.
-setPDFNetworkStreamFactory(function (params) {
-  return new PDFNodeStream(params);
-});
